Update TLW icon when the DPI changes in wxMSW
While we selected the icon of the size most appropriate for the current
DPI when the window was created, we didn't update the icon later if the
DPI changed as we should, so do it now.

Note that we abuse the existing MSWUpdateFontOnDPIChange() to make it
simple to backport this fix to 3.2 without breaking the ABI.

See #22807, #22831.

(cherry picked from commit 2d3da75cf2e3f45d71da12e5f249f73f66a9310f)