Fix type of values passed to TIFFSetField()
This vararg function accepts any kind of values, without any checks, but
actually needs double for TIFFTAG_[XY]RESOLUTION and not int that we
passed to it, and which could result in the double values actually
having a NaN value of the word on the stack before the int happened to
have just the wrong value and which, in turn, resulted in a failure to
save the TIFF file later as libtiff checks that the values are valid,
i.e. positive and not-NaN. Note that we really need to pass the values
as doubles, even if they're stored as floats internally and returned as
such by TIFFGetField().

Also fix the types for all the other calls of this function, even if
this hasn't been observed to result in any problems so far.

See #23048, #23056.

(cherry picked from commit af923360abf278bf8a488238a85547d06344d4e4)