Fix wrongly recreating wxComboBox in disabled state in wxMSW
When recreating a wxComboBox, as sometimes needs to be done when adding
a new bitmap to wxBitmapComboBox, for example, we should only disable it
if the old combobox itself was disabled, not if (any of) its parent(s)
were disabled, but IsEnabled() returned true even in this case.

Use IsThisEnabled() instead to ensure that we only disable the new
combobox if necessary.

See #23132.

(cherry picked from commit 8e604926c3a726c93b2df2c162be57b94a10cb61)