Disable using [[deprecated]] attribute when not using C++11
At least with g++ 10, but presumably with other versions too, the
compiler still predefines __has_cpp_attribute(deprecated) even when
using it with -std=c++98 or similar option, but it does not accept this
attribute unless it's actually using C++11 or higher, resulting in
compilation errors.

Simply avoid defining wxHAS_DEPRECATED_ATTR when not using C++11 to fix
this.