Fix using pkg-config when cross-compiling to Linux targets
Restrict the hack disabling the use of build system pkg-config files
added in ba3037e888 (Disable use of build system pkg-config files when
cross-compiling, 2020-12-13) to the case of using non-host-prefixed
pkg-config script and cross-compiling to non-Linux targets, as we're
very unlikely to have the correct .pc files installed on the build
system, but do use pkg-config in the other cases, as we do want to use
the system-provided .pc files then.

This is a combined cherry-pick of d6ddfe87e1 (Disable build system
pkg-config when cross-compiling to non-Linux, 2022-11-30), 7899850496
(Do use host-specific pkg-config when cross-compiling, 2022-12-15) and
b0056f8ce5 (Re-re-enable using pkg-config when targeting Linux systems,
2023-01-25) from master.

See #22886, #23037, #23171.