Don't look for C++11 keywords in the comments
Avoid flagging the use of "override" in a comment as a C++11 keyword by
stripping the comments before checking for the keywords forbidden in
this branch.

This is not 100% correct and doesn't deal with multiline comments at
all, but we use the latter ones rarely and it's not worth the effort to
add support for them here.