Fix setting locale to wxLANGUAGE_UKRAINIAN
Use GetCanonicalWithRegion() (instead of LocaleTag)  in the Windows
implementation of method wxLanguageInfo::TrySetLocale() -- as is done in
the non-Windows implementation of this method -- and to ensure that we
use the correct "uk-UA" for Ukrainian instead of the wrong (because
conflicting with the UK English) "uk".

See #23269.

Closes #23210.

(cherry picked from commit a0be352a3f95253dde0de952fb41f7157ebd1840)