Improve code in gtk_window_key_press_callback()
Don't use isalpha() in it as this was potentially incorrect, both
because the key code could be not in the supported range and because
isalpha() might return unexpected result in non C locale, and also
inconsistent with AdjustCharEventKeyCodes().

Just check that the code is in a..z or A..Z range explicitly instead.

See #23379.

(cherry picked from commit 4799785b6dfc2d4490598b447c99435105b8ebba)