Add wxUILocale::GetMonthName() and GetWeekDayName()
Unlike in master, only the variants for the date/time formatting context
are returned to reduce the number of changes in this branch. This is
wrong for standalone month names like in the generic calendar control.
However, in the prior implementation that was also the case.

Use a horrible hack to avoid adding new virtual functions to
wxUILocaleImpl class which, although private, still somehow results in
ABI breakage.

See #23556, #23581.

Co-Authored-By: Vadim Zeitlin <vadim@wxwidgets.org>