Fix wxGLCanvasEGL::IsShownOnScreen() under Wayland
Don't use m_readyToDraw in it, as the canvas is still shown even when
it's not ready to draw yet, and use m_wlSubsurface instead as it is
destroyed when the window is unmapped.

See #23899.

(cherry picked from commit cf44a87ec62f9cefb428b8d6c002c2ee342dc4ec)