Document that wxLogInterposerTemp shouldn't be used
Mark it as deprecated in the documentation without deprecating it
formally in the code, as it seems to be used in some existing code and
there is no urgency to remove it, but using it in new code still doesn't
make any sense.

See #23945.

(cherry picked from commit 543fd3fabcb6d88a6e220520c41501c3f961dd55)