Prefer EGL configurations without alpha if it wasn't requested
At least Mesa Iris on Wayland is known to put RGBA5551 and RGBA4444
formats at the start of the config list, even if we didn't request
alpha, which is unexpected, so filter them out manually in this case to
ensure we get the expected RGB888, as with the other drivers.

See #24395, #24397.

Co-authored-by: Vadim Zeitlin <vadim@wxwidgets.org>

(cherry picked from commit ff90bd59e96940a1c3805f618fd8b08e0a162693)