Disable -Wdepracated-copy-dtor in Circle CI builds
The warning was fixed in fc35ad92bb (Remove unnecessary empty
destructors, 2024-01-26) in master but this change is not
ABI-compatible, so can't be backported to 3.2 -- just disable the
warning instead.

For now only do it in Circle CI workflow, but we might want to do it in
the wx headers themselves in the future.