Fix crash when handling diagonal scroll events in wxOSX
Commit a0b4c5f14b (Generate correct events for diagonal trackpad
scrolling under Mac, 2024-06-09) introduced a crash due to the use of a
reference which became dangling after resizing the vector, which could
result in a reallocation.

Fix this by replacing the reference with a pointer and resetting this
pointer after resize.

See #24516, #24601, #24649.

(cherry picked from commit 35040847c208682ec3d0cac559c06329446573fc)