Improve layout of wizard pages using non-default fonts in wxGTK
Remove the ugly hack with wizard page visibility added back in
0a9ed8f14e (trick the sizer code into thinking that the wizard pages are
shown so that it takes them into account for the layout, even if they're
not really shown, 2006-02-12) and which broke the attempts to work
around GTK style caching in wxGTK -- if we don't pretend that the page
is shown when it really isn't, the existing workarounds seem to work.

And we can just use wxRESERVE_SPACE_EVEN_IF_HIDDEN to achieve the same
effect with less code.

See #19053.

(cherry picked from commit 7e14fda22c3a25379683e104f339206ca1251d2e)