Ellipsize info bar text instead of truncating it
If there is not enough space to show the text in full, prefer to
ellipsize it in the middle instead of just truncating it.

Also show some text too long to fit into the window in the dialogs
sample to demonstrate that ellipsization really works.

(cherry picked from commit 3f5bbe541b99725bc6be1c9fe6d88771a1cfe4df)