Don't remove the old file before renaming new one to it
Rename replaces the existing file, so it is not necessary to remove it
before calling it and this is actually dangerous and may result in data
loss if renaming fails, as the original file contents is just lost then.

See #25088.

(cherry picked from commit 6f861d376be2938285d6ec6fca710b5514e8459d)