Generate key codes consistent with wxMSW for VK_OEM_xxx in wxGTK
These keys should generate the key code corresponding to their label on
the physical keyboard if it's a Latin-1 character both because this is
how wxMSW behaves and because this is likely what the user expects.

Do this by using the Unicode character corresponding to one of these
keys if it's a Latin-1 one, just as we already did it for the letters.

(cherry picked from commit 20a98a0842d938129be863c6eda7cbb13548c231)