Call OnExceptionInMainLoop() in wxGTK if idle event handler throws
As idle events are (only) generated while an event loop is running, it
makes sense to call OnExceptionInMainLoop() if a handler for such an
event throws an exception.

This was already the case in wxMSW, where the exception was caught in
wxEventLoopManual code from where ProcessIdle() was called, but not
wxGTK, where it was called directly from an extern "C" callback.
And while catching the exception earlier does no harm in wxMSW, it fixes
the problem in wxGTK.

See #25312.

(cherry picked from commit ea351362a34531f7828701e5fce401f5db78833d)