Fix defining ELF symbols versions when flavour is used
Unfortunately the changes of 362ced55ff (Fix ABI breakage for versioned
symbols in 3.2.7, 2025-04-14) themselves broke another use case, as they
didn't take into account the fact that ELF version includes the library
flavour and hardcoded the version without it in elfversion.h.

Fix this by defining wxELF_VERSION_PREFIX in configure and using it in
this header to use proper symver value even when flavour is not empty.

See #25327.

Closes #25436.