Fix broken wxMenuItem bitmap setting on macOS
This reverts most of 59bba56 and b2a3645 (with small bits of f096073)
and replaces the error-prone implementation with a simple and
straightforward one.

On macOS, NSImage is a direct equivalent of wxBitmapBundle and looking
up and converting a specific-scale bitmap when the menu is attached is
unnecessary busywork.

Instead, switch wxMenuItemImpl to use wxBitmapBundle too. This is
private API, so there are no compatibility concerns.

Among the bugs in the previous implementation were:
- ignoring bitmaps set in the global Mac menu
- ignoring bitmaps set in the Mac-specific application menu
- loosing compatibility with NSImage-backed wxBitmap and rendering them
  without appearance awareness and in low resolution in the global menu

See #25744.