Optimize generic wxHeaderCtrl drawing and work around wxOSX bug
Don't draw all the columns, there is no need to do it for those scrolled
out of view.

As a side effect, don't rely on wxDC::SetDeviceOrigin() for the
coordinates translation but do it ourselves in the drawing code: this is
pretty trivial in this case and avoids a bug in wxOSX which doesn't seem
to honour device origin in its wxRendererNative implementation, somehow.

See #26014, #26016.

(cherry picked from commit 411430b835274bd234947adae467f2126210d6dd)